chore: move getConversationQualifiedMembers out of the proteus domain
