+++ packages/core/src/messagingProtocols/proteus/ProteusService/Utility/Recipients.ts
@@ -21,9 +21,9 @@ import {APIClient} from '@wireapp/api-client/lib/APIClient';
 import {QualifiedUserClients, UserClients} from '@wireapp/api-client/lib/conversation';
 import {QualifiedId, QualifiedUserPreKeyBundleMap, UserPreKeyBundleMap} from '@wireapp/api-client/lib/user';
 
-import {getConversationQualifiedMembers} from './getConversationQualifiedMembers';
-import {getPreKeyBundleMap, getQualifiedPreKeyBundle} from './PreKeyBundle';
+import {getQualifiedPreKeyBundle, getPreKeyBundleMap} from './PreKeyBundle';
 
+import {getConversationQualifiedMembers} from '../../../../conversation/ConversationService/Utility/getConversationQualifiedMembers';
 import {isQualifiedUserClients, isUserClients} from '../../../../util';
 
 interface GetRecipientsForConversationQualifiedParams {
