+++ packages/core/src/messagingProtocols/proteus/ProteusService/Utility/index.ts
@@ -18,6 +18,6 @@
  */
 
 export * from './PreKeyBundle';
-export * from './getConversationQualifiedMembers';
+export * from '../../../../conversation/ConversationService/Utility/getConversationQualifiedMembers';
 export * from './UserIds';
 export * from './Recipients';
